+++ swim/scripts/constants.js
@@ -20,3 +20,36 @@ export const NUM_MORPHS_MAX = 100
 
 export const ITEM_CONFIG_WINDOW_WIDTH = 500;
 export const ITEM_CONFIG_WINDOW_HEIGHT = 400;
+
+export function get_compendiums_list(type) {
+    const options = []
+
+    // Get all packs:
+    for (let pack of game.packs) {
+        if (pack.documentName.toLowerCase() === type.toLowerCase() || type.toLowerCase() === 'any') {
+            let option = {
+                value: pack.collection,
+                name: pack.title
+            };
+            options.push(option)
+        }
+    }
+
+    // Sort packs by name:
+    options.sort((a, b) => {
+        const nameA = a.name.toUpperCase() // Convert to uppercase for case-insensitive sorting
+        const nameB = b.name.toUpperCase()
+        if (nameA < nameB) {
+            return -1
+        }
+        if (nameA > nameB) {
+            return 1
+        }
+        return 0 // Names are equal
+    })
+
+    // Add the "none" object at the first position:
+    options.unshift({ value: 'none', name: game.i18n.localize('SWIM.none')})
+
+    return options
+}